[msbuild] Merge the UnpackLibraryResources[TaskBase] classes.
Also enable nullability and fix the resulting issues.
